RN-685: Moved mocks out of __tests__ and into main source code
