+++ packages/api-client/src/connections/mocks/index.ts
@@ -7,4 +7,3 @@ export { MockAuthApi } from './MockAuthApi';
 export { MockCentralApi } from './MockCentralApi';
 export { MockEntityApi } from './MockEntityApi';
 export { MockReportApi } from './MockReportApi';
-export { MockTupaiaApiClient } from './MockTupaiaApiClient';
